Provide bare bones to start building a Saga
Add an Order core api, containing the OrderPlacedEvent and a
Confirm/Reject Order command. Second, an OrderCommandHandler should be
introduced the logs the outcome, to show the Saga worked. Then add an
OrderCommandHandler bean and an EventGateway bean. The latter can be
used in the GiftCardController to spoof an OrderPlacedEvent

#demo-saga